small fixes

+++ render_flac_files.py
@@ -3,6 +3,7 @@ import os
 import time
 import utils
 import argparse
+import pretty_midi as pm
 
 parser = argparse.ArgumentParser()
 parser.add_argument('piano', type=str, help='specify the piano model in use')
@@ -17,17 +18,33 @@ utils.create_path(flac_path)
 
 processes = []
 
-for project in utils.get_files_by_suffix(projects_path, '.rpp'):
+count = 0
+for i, project in enumerate(utils.get_files_by_suffix(projects_path, '.rpp')):
     if os.path.exists(os.path.join(flac_path, project[:-4]+'.flac')):
         continue
-    print(project)
+    print(i, project)
     project_file = os.path.join(projects_path, project)
     p = subprocess.Popen(' '.join([reaper, '-renderproject', project_file]))
-    time.sleep(20)
+    midi_data = pm.PrettyMIDI('data/dataset-temp/'+project[:-3]+'mid')
+    music_length = midi_data.get_end_time()
+    print('music length:', music_length)
+    time.sleep(music_length / 25)
     processes.append(p)
+    
+    if count % 5 == 4:
+        print('rest for a while...')
+        time.sleep(10)
+        for p in processes:
+            p.kill()
+        print('continue rendering~~~')
+    count += 1
 
 print('rendering finished!')
 
 time.sleep(10)
-for p in processes:
-    p.kill()
+try:
+    for p in processes:
+        p.kill()
+    print('exit')
+except:
+    print('exit')
